[plugin.video.waipu] 0.3.3
