+++ plugin.video.waipu.tv/addon.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="utf-8" standalone="yes"?>
-<addon id="plugin.video.waipu.tv" name="Waipu.tv / O2 TV" version="0.3.1" provider-name="flubshi, MiRo">
+<addon id="plugin.video.waipu.tv" name="Waipu.tv / O2 TV" version="0.3.3" provider-name="flubshi, MiRo">
 <requires>
   <import addon="xbmc.python" version="2.25.0"/>
   <import addon="script.module.requests" version="2.12.4"/>
   <import addon="script.module.inputstreamhelper" version="0.3.3"/>
   <import addon="script.module.dateutil" version="2.7.3" />
   <import addon="script.module.mechanize" version="0.2.6" />
+  <import addon="script.module.routing" version="0.2.0"/>
 </requires>
 <extension point="xbmc.python.pluginsource" library="main.py">
   <provides>video</provides>
@@ -30,6 +31,8 @@
     <screenshot>resources/screenshots/screenshot-03.jpg</screenshot>
   </assets>
   <news>
+- 0.3.3 fix internal imports
+- 0.3.2 Implement VoD
 - 0.3.1 Adjust mechanize dependency to keep Leia compatible
 - 0.3.0 Add python3 support
 - 0.2.9 Display and play active recordings
